+++ app/src/component/shapes.tsx
@@ -5,10 +5,9 @@ export interface ShapesProps {
     shapes: Shape[];
 }
 export function Shapes({ shapes }: ShapesProps) {
-    {
-        return (shapes.length > 0 && shapes
-            .map((points, idx) => (
-                <Polyline key={idx} positions={points} color="blue" weight={4} opacity={0.7} />
-            )))
-    }
+    return (shapes.length > 0 && shapes
+        .map((points, idx) => (
+            <Polyline key={idx} positions={points} color="blue" weight={4} opacity={0.7} />
+        )))
+
 }
